+++ src/defaults/get-bump.js
@@ -1,10 +1,10 @@
-export default function(commits) {
-  let bump = 'patch';
-  commits.forEach(function(commit) {
+export default function (commits) {
+  let bump = "patch";
+  commits.forEach(function (commit) {
     if (commit.breaking) {
-      bump = 'major';
-    } else if (commit.type === 'feat' && bump === 'patch') {
-      bump = 'minor';
+      bump = "major";
+    } else if (commit.type === "feat" && bump === "patch") {
+      bump = "minor";
     }
   });
   return bump;