+++ src/plugins/bump-package-json/index.js
@@ -1,18 +1,18 @@
-import {writeFileSync, readFileSync} from 'fs';
-import {resolve} from 'path';
+import { writeFileSync, readFileSync } from "fs";
+import { resolve } from "path";
 
 export default function installPlugin(release) {
-  release.phases.start.before('commit', {
-    name: 'bumpPackageJson',
+  release.phases.start.before("commit", {
+    name: "bumpPackageJson",
     run(context) {
-      let packageJsonPath = resolve(process.cwd(), 'package.json');
+      let packageJsonPath = resolve(process.cwd(), "package.json");
       let packageJson = JSON.parse(readFileSync(packageJsonPath));
       packageJson.version = context.nextVersion;
 
       writeFileSync(
         packageJsonPath,
-        JSON.stringify(packageJson, null, '  ') + '\n'
+        JSON.stringify(packageJson, null, "  ") + "\n"
       );
-    }
+    },
   });
 }